RL-Separación del Middleware
