+++ src/routes/user.routes.js
@@ -2,35 +2,7 @@ import express from "express";
 const routes = express.Router();
 import db from "../models/index.js";
 const { Users } = db;
-
-//Middleware para localizar el usuario por su ID:
-const getUser = async (req, res, next) => {
-  let user;
-  const { id } = req.params;
-
-  if (isNaN(id) || parseInt(id) <= 0) {
-    return res.status(400).json({
-      message: "El Id dado no es valido.",
-    });
-  }
-
-  try {
-    user = await Users.findByPk(id);
-
-    if (!user) {
-      return res.status(404).json({
-        message: "Usuario no encontrado.",
-      });
-    }
-  } catch (error) {
-    return res.status(500).json({
-      message: error.message,
-    });
-  }
-
-  res.user = user;
-  next();
-};
+import getUser from "../middleware/getUser.middleware.js";
 
 //Rutas de las operaciones del CRUD Users:
 //Mostrar todos los usuarios:
